ref(data-collection)!: Move option to top level
Move dataCollection from experimental options to Options and decode it from\nthe top-level configuration dictionary.\n\nFixes #8368
