+++ CHANGELOG_V10.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Features
+
+- Add `options.dataCollection` to configure data collection behaviour (#8448)
+- Add dictionary initialization for `options.dataCollection` (#8371)
+
 ### Breaking Changes
 
 - Remove Objective-C `@objc` attributes from SentrySDK (#8308)
