+++ Sources/SentryObjC/Public/SentryObjCExperimentalOptions.h
@@ -1,8 +1,4 @@
 #import <Foundation/Foundation.h>
-#if SDK_V10
-@class SentryObjCDataCollectionOptions;
-#endif
-
 NS_ASSUME_NONNULL_BEGIN
 
 /// Options for experimental features that are subject to change or may be removed in future
@@ -24,11 +20,6 @@ NS_ASSUME_NONNULL_BEGIN
 
 /// When enabled, the SDK uses a more efficient mechanism for detecting watchdog terminations.
 @property (nonatomic) BOOL enableWatchdogTerminationsV2;
-
-#if SDK_V10
-/// Configuration for what data the SDK collects automatically.
-@property (nonatomic, strong) SentryObjCDataCollectionOptions *dataCollection;
-#endif // SDK_V10
 
 /// Initializes experimental options with default values.
 - (instancetype)init;
