+++ Sources/Swift/Options+Dictionary.swift
@@ -300,9 +300,11 @@ extension Options {
             self.spotlightUrl = spotlightUrl
         }
 
-        if let experimental = dictionary["experimental"] as? [String: Any] {
-            self.experimental.populateFrom(dict: experimental)
+        #if SDK_V10
+        if let dataCollection = dictionary["dataCollection"] as? [String: Any] {
+            self.dataCollection = SentryDataCollection.Options(dictionary: dataCollection)
         }
+        #endif // SDK_V10
     }
 
     private func boolValue(_ value: Any?) -> Bool? {
