+++ Sources/Swift/Options.swift
@@ -702,6 +702,16 @@
     // swiftlint:disable:next missing_docs
     @_spi(Private) @objc public static let defaultEnvironment = "production"
 
+#if SDK_V10
+    // MARK: - Data Collection
+
+    /// Configuration for what data the SDK collects automatically (e.g. headers, cookies, query params, bodies, user identity).
+    ///
+    /// - Important: Replaces `sendDefaultPii` with granular per-category control.
+    /// - SeeAlso: https://docs.sentry.io/platforms/apple/configuration/options/#dataCollection
+    public var dataCollection = SentryDataCollection.Options()
+#endif // SDK_V10
+
     // MARK: - Integration: Metrics
 
     /// When enabled, the SDK sends metrics to Sentry. Metrics can be captured using the ``SentrySDK/metrics``
