+++ Tests/SentryObjCTests/SentryObjCExperimentalOptionsTests.m
@@ -76,120 +76,4 @@
     XCTAssertTrue(options.enableWatchdogTerminationsV2);
 }
 
-#pragma mark - dataCollection
-
-- (void)testDataCollection_whenDefault_shouldReturnNotNil
-{
-#if !SDK_V10
-    XCTSkip(@"Test skipped for SDK_V10");
-#else
-    // -- Arrange --
-    SentryObjCExperimentalOptions *options = [[SentryObjCExperimentalOptions alloc] init];
-
-    // -- Assert --
-    XCTAssertNotNil(options.dataCollection);
-#endif
-}
-
-- (void)testDataCollection_whenDefault_shouldHaveSpecDefaults
-{
-#if !SDK_V10
-    XCTSkip(@"Test skipped for SDK_V10");
-#else
-    // -- Arrange --
-    SentryObjCExperimentalOptions *options = [[SentryObjCExperimentalOptions alloc] init];
-
-    // -- Assert --
-    XCTAssertTrue(options.dataCollection.userInfo);
-    XCTAssertEqual(options.dataCollection.httpBodies, SentryObjCDataCollectionHttpBodyTypeAll);
-    XCTAssertTrue(options.dataCollection.stackFrameVariables);
-    XCTAssertEqual(options.dataCollection.frameContextLines, 5u);
-#endif
-}
-
-- (void)testDataCollection_whenSet_shouldReturnNewValue
-{
-#if !SDK_V10
-    XCTSkip(@"Test skipped for SDK_V10");
-#else
-    // -- Arrange --
-    SentryObjCExperimentalOptions *options = [[SentryObjCExperimentalOptions alloc] init];
-    SentryObjCDataCollectionOptions *dc = [[SentryObjCDataCollectionOptions alloc] init];
-    dc.userInfo = NO;
-
-    // -- Act --
-    options.dataCollection = dc;
-
-    // -- Assert --
-    XCTAssertFalse(options.dataCollection.userInfo);
-#endif
-}
-
-- (void)testDataCollection_whenMutatedInPlace_shouldPropagateToParent
-{
-#if !SDK_V10
-    XCTSkip(@"Test skipped for SDK_V10");
-#else
-    // -- Arrange --
-    SentryObjCExperimentalOptions *options = [[SentryObjCExperimentalOptions alloc] init];
-    XCTAssertTrue(options.dataCollection.userInfo);
-
-    // -- Act --
-    options.dataCollection.userInfo = NO;
-
-    // -- Assert --
-    XCTAssertFalse(options.dataCollection.userInfo);
-#endif
-}
-
-- (void)testDataCollection_whenSubPropertyMutatedInPlace_shouldPropagateToParent
-{
-#if !SDK_V10
-    XCTSkip(@"Test skipped for SDK_V10");
-#else
-    // -- Arrange --
-    SentryObjCExperimentalOptions *options = [[SentryObjCExperimentalOptions alloc] init];
-
-    // -- Act --
-    options.dataCollection.database.queryParams = NO;
-
-    // -- Assert --
-    XCTAssertFalse(options.dataCollection.database.queryParams);
-#endif
-}
-
-- (void)testDataCollection_whenGraphqlMutatedInPlace_shouldPropagateToParent
-{
-#if !SDK_V10
-    XCTSkip(@"Test skipped for SDK_V10");
-#else
-    // -- Arrange --
-    SentryObjCExperimentalOptions *options = [[SentryObjCExperimentalOptions alloc] init];
-
-    // -- Act --
-    options.dataCollection.graphql.document = NO;
-
-    // -- Assert --
-    XCTAssertFalse(options.dataCollection.graphql.document);
-#endif
-}
-
-- (void)testDataCollection_whenHttpHeadersMutatedInPlace_shouldPropagateToParent
-{
-#if !SDK_V10
-    XCTSkip(@"Test skipped for SDK_V10");
-#else
-    // -- Arrange --
-    SentryObjCExperimentalOptions *options = [[SentryObjCExperimentalOptions alloc] init];
-
-    // -- Act --
-    options.dataCollection.httpHeaders.request =
-        [SentryObjCDataCollectionKeyValueCollectionBehavior off];
-
-    // -- Assert --
-    XCTAssertEqual(options.dataCollection.httpHeaders.request.mode,
-        SentryObjCDataCollectionKeyValueCollectionModeOff);
-#endif
-}
-
 @end
