+++ Tests/SentryTests/DataCollection/SentryDataCollectionOptionsTests.swift
@@ -506,116 +506,6 @@ class SentryDataCollectionOptionsTests: XCTestCase {
         #endif
     }
 
-    // MARK: - Options Integration
-
-    func testOptions_dataCollection_hasSpecDefaults() throws {
-        #if !SDK_V10
-        throw XCTSkip("Test skipped for SDK_V10")
-        #else
-        // -- Act --
-        let options = Options()
-
-        // -- Assert --
-        XCTAssertEqual(options.experimental.dataCollection, SentryDataCollection.Options())
-        #endif
-    }
-
-    func testOptions_dataCollection_canBeSet() throws {
-        #if !SDK_V10
-        throw XCTSkip("Test skipped for SDK_V10")
-        #else
-        // -- Arrange --
-        let options = Options()
-
-        // -- Act --
-        options.experimental.dataCollection = SentryDataCollection.Options(userInfo: false)
-
-        // -- Assert --
-        XCTAssertFalse(options.experimental.dataCollection.userInfo)
-        #endif
-    }
-
-    func testOptionsWithDictionary_whenExperimentalDataCollectionIsPresent_shouldSetDataCollection() throws {
-        #if !SDK_V10
-        throw XCTSkip("Test skipped for SDK_V10")
-        #else
-        // -- Arrange --
-        let dictionary: [String: Any] = [
-            "dsn": "https://username:password@sentry.io/1",
-            "experimental": [
-                "dataCollection": [
-                    "userInfo": false
-                ]
-            ]
-        ]
-
-        // -- Act --
-        let options = try Options(dictionary: dictionary)
-
-        // -- Assert --
-        XCTAssertFalse(options.experimental.dataCollection.userInfo)
-        XCTAssertEqual(options.experimental.dataCollection.cookies, SentryDataCollection.Options().cookies)
-        #endif
-    }
-
-    func testOptionsWithDictionary_whenExperimentalDataCollectionIsAbsent_shouldUseDefault() throws {
-        #if !SDK_V10
-        throw XCTSkip("Test skipped for SDK_V10")
-        #else
-        // -- Arrange --
-        let dictionary: [String: Any] = [
-            "dsn": "https://username:password@sentry.io/1",
-            "experimental": [:]
-        ]
-
-        // -- Act --
-        let options = try Options(dictionary: dictionary)
-
-        // -- Assert --
-        XCTAssertEqual(options.experimental.dataCollection, SentryDataCollection.Options())
-        #endif
-    }
-
-    func testOptionsWithDictionary_whenExperimentalDataCollectionHasWrongType_shouldUseDefault() throws {
-        #if !SDK_V10
-        throw XCTSkip("Test skipped for SDK_V10")
-        #else
-        // -- Arrange --
-        let dictionary: [String: Any] = [
-            "dsn": "https://username:password@sentry.io/1",
-            "experimental": [
-                "dataCollection": "off"
-            ]
-        ]
-
-        // -- Act --
-        let options = try Options(dictionary: dictionary)
-
-        // -- Assert --
-        XCTAssertEqual(options.experimental.dataCollection, SentryDataCollection.Options())
-        #endif
-    }
-
-    func testOptionsWithDictionary_whenExperimentalDataCollectionIsNSNull_shouldUseDefault() throws {
-        #if !SDK_V10
-        throw XCTSkip("Test skipped for SDK_V10")
-        #else
-        // -- Arrange --
-        let dictionary: [String: Any] = [
-            "dsn": "https://username:password@sentry.io/1",
-            "experimental": [
-                "dataCollection": NSNull()
-            ]
-        ]
-
-        // -- Act --
-        let options = try Options(dictionary: dictionary)
-
-        // -- Assert --
-        XCTAssertEqual(options.experimental.dataCollection, SentryDataCollection.Options())
-        #endif
-    }
-
     // MARK: - Equatable
 
     func testEquatable_defaultInstances_shouldBeEqual() throws {
